+++ tests/TestCase/Controller/ThumbsControllerTest.php
@@ -57,7 +57,7 @@ class ThumbsControllerTest extends IntegrationTestCase
         parent::tearDown();
 
         //Deletes all thumbnails
-        foreach (glob(Configure::read('Thumbs.target') . DS . '*') as $file) {
+        foreach (glob(Configure::read(THUMBER . '.target') . DS . '*') as $file) {
             unlink($file);
         }
     }