+++ tests/bootstrap.php
@@ -97,12 +97,7 @@ Cache::setConfig([
     ],
 ]);
 
-Configure::write('Thumbs.driver', 'imagick');
-Configure::write('Thumbs.target', TMP . 'thumbs');
 Configure::write('Thumbs.comparingDir', TESTS . DS . 'comparing_files' . DS . 'imagick' . DS);
-
-//@codingStandardsIgnoreLine
-@mkdir(Configure::read('Thumbs.target'));
 
 Plugin::load('Thumber', [
     'bootstrap' => true,